image cache: split the repository out of the cache
`caches/image_cache.c` was two things wearing one name: an LRU of dt_image_t
structs with refcounting and per-entry locking, and the only code in the tree
that knows the shape of a `main.images` row. 107 of its lines were SQL. The
second is not a cache, it is a repository, and keeping them together meant every
consumer of the image cache also had a translation unit that could reach the
database.

  caches/image_cache.c          548 lines,   0 sqlite refs
  database/image_repository.c   320 lines, 103 sqlite refs, NO upward includes

`src/database` starts here, at layer 1. The repository holds the load statement,
the write statement, the history-hash upsert and their prepared-statement cache;
the cache calls dt_image_repository_load() and _store() and knows nothing about
SQL. The SQL bodies moved verbatim -- the split cannot change what is written.

**dt_image_from_stmt() was not a row mapper.** Past the columns it computed
derived fields -- rating, monochrome and HDR predicates, makermodel, and an
extension cross-check that logs stale stored flags -- and those need
dt_imageio_get_type_from_extension() from imageio/ (layer 6) and DT_VIEW_REJECT
from views/ (layer 7). A "database row to struct" function reaching two layers up
is a large part of why this code could never leave common/. The columns stay in
the repository; the derivation becomes dt_image_derive_fields() in the cache, and
both producers of a dt_image_t from a row call it -- the repository's own load,
and gui/dtgtk/thumbtable.c's bulk query, which previously got it for free and now
says so.

**dt_view_image_over_t moves from views/view.h to common/image.h.** It is the
value stored in dt_image_t.rating (0-5 stars, 6 = rejected), and it was read 8
times from common/, 3 from caches/ and 5 from gui/ against exactly ONE use in
views/ -- an enum declared at the top layer and consumed from the bottom. It was
the only reason common/ratings.c and caches/image_cache.c included the view
layer at all; both drop the include.

Layering violations 206 -> 205, and the baseline is lowered in this commit.
Naming views/view.h explicitly first pushed it to 207, which is the honest
intermediate: the dependency was always there, arriving transitively.

Exported pixels identical to master on a raw, GPU and CPU. Four build
configurations.
